+++ script/test.sol
@@ -43,7 +43,7 @@ contract KintoDeployTestWalletScript is AASetup, KYCSignature {
         if (!_kintoID.isKYC(recipientWallet)) {
             IKintoID.SignatureData memory sigdata =
                 _auxCreateSignature(_kintoID, recipientWallet, recipientWallet, recipientKey, block.timestamp + 1000);
-            uint8[] memory traits = new uint8[](0);
+            uint16[] memory traits = new uint16[](0);
             _kintoID.mintIndividualKyc(sigdata, traits);
         }
