+++ test/helpers/AATestScaffolding.sol
@@ -180,13 +180,13 @@ abstract contract AATestScaffolding is KYCSignature {
 
         IKintoID.SignatureData memory sigdata =
             _auxCreateSignature(_kintoIDv1, _account, _account, _accountPk, block.timestamp + 1000);
-        uint8[] memory traits = new uint8[](0);
+        uint16[] memory traits = new uint16[](0);
         _kintoIDv1.mintIndividualKyc(sigdata, traits);
 
         vm.stopPrank();
     }
 
-    function approveKYC(address _kycProvider, address _account, uint256 _accountPk, uint8[] memory traits) public {
+    function approveKYC(address _kycProvider, address _account, uint256 _accountPk, uint16[] memory traits) public {
         vm.startPrank(_kycProvider);
 
         IKintoID.SignatureData memory sigdata =
